+++ src/interfaces/ISpokeConfigurator.sol
@@ -165,7 +165,7 @@ interface ISpokeConfigurator {
   function updateDynamicReserveConfig(
     address spoke,
     uint256 reserveId,
-    uint256 configKey,
+    uint16 configKey,
     DataTypes.DynamicReserveConfig calldata dynamicConfig
   ) external;
